Splitting the content that used to be solely in main.cpp into different files.
