+++ Container_Classes/src/ContainerAdaptors.hpp
@@ -1,0 +1,12 @@
+#ifndef CONTAINER_ADAPTORS_HPP
+#define CONTAINER_ADAPTORS_HPP
+
+/******** Function prototypes *********/
+
+void containerAdaptorsStacks(void);
+void containerAdaptorsQueues(void);
+void containerAdaptorsHeaps(void);
+
+/**************************************/
+
+#endif
